// enums/sampling-fields.ts
export enum SamplingFieldsEnum {
  TRAFFIC = 'Tráfico de CLIENTES',
  ETHNICITY = 'Grupo Étnico CLIENTES',
  AGE_RANGE = 'Rango de Edad CLIENTES',
  GENDER = 'Género CLIENTES',
  PURCHASE_INTENTION = '¿Qué FACTORES influyeron HOY en la DECISIÓN de COMPRA del Producto?',
  CONSUMPTION_MOMENT = '¿En que MOMENTO CONSUME el Producto usualmente?',
  NET_PROMOTER_SCORE = '¿ Que tanto RECOMENDARÍA la Marca?  (Del 1 al 10, donde el 10 es que la Recomendaría Ampliamente)',
  FOLLOW_UP = '¿Siguió a la Marca en Redes Sociales? (Posterior a que el (la) Promotor(a) le haya sugerido hacerlo)',
  CLIENTS_COMMENTS = 'Comentarios de los CLIENTES',
  PROMOTOR_COMMENTS = 'Comentarios de el (la) PROMOTOR (A))',
}